Refactor config and database utilities
