+++ app.py
@@ -14,7 +14,6 @@ import jwt
 import urllib.parse
 from flask import (
     Flask,
-    g,
     render_template,
     request,
     redirect,
@@ -26,16 +25,29 @@ from flask import (
     Response,
 )
 from markupsafe import escape
+from config import Config
+from database import (
+    get_db,
+    close_connection,
+    query_db,
+    execute_db,
+    init_db,
+    ensure_schema,
+    create_new_db,
+    _sanitize_db_name,
+    _sanitize_export_name,
+)
 
 app = Flask(__name__)
+app.config.from_object(Config)
 app.jwt = jwt
-# Allow overriding the startup database via environment variable
-env_db = os.environ.get('RETRORECON_DB')
+env_db = app.config.get('DB_ENV')
 if env_db:
     app.config['DATABASE'] = env_db if os.path.isabs(env_db) else os.path.join(app.root_path, env_db)
 else:
     app.config['DATABASE'] = None
-app.secret_key = 'CHANGE_THIS_TO_A_RANDOM_SECRET_KEY'
+app.secret_key = app.config.get('SECRET_KEY', 'CHANGE_THIS_TO_A_RANDOM_SECRET_KEY')
+app.teardown_appcontext(close_connection)
 ITEMS_PER_PAGE = 10  # default results per page
 ITEMS_PER_PAGE_OPTIONS = [5, 10, 15, 20, 25]
 TEXT_TOOLS_LIMIT = 64 * 1024  # 64 KB limit for text transformations
@@ -363,27 +375,6 @@ def take_screenshot(url: str, user_agent: str = '', spoof_referrer: bool = False
         except Exception:
             return base64.b64decode("iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAQAAAC1HAwCAAAAC0lEQVR42mP8z8AAPAIB+AWd1QAAAABJRU5ErkJggg==")
 
-def init_db() -> None:
-    """Initialize the database using the schema.sql file."""
-    schema_path = os.path.join(app.root_path, 'db', 'schema.sql')
-    if not os.path.exists(schema_path):
-        raise FileNotFoundError("schema.sql not found")
-    with open(schema_path, "r", encoding="utf-8") as f:
-        sql = f.read()
-
-    conn = sqlite3.connect(app.config['DATABASE'])
-    for statement in sql.split(';'):
-        stmt = statement.strip()
-        if stmt.upper().startswith("CREATE TABLE IF NOT EXISTS"):
-            conn.execute(stmt)
-    conn.commit()
-    conn.close()
-
-def ensure_schema() -> None:
-    """Apply ``schema.sql`` to an existing database if tables are missing."""
-    if os.path.exists(app.config['DATABASE']):
-        init_db()
-
 def load_demo_data() -> None:
     """Populate the database with entries from ``DEMO_DATA_FILE`` if present."""
     if not os.path.exists(DEMO_DATA_FILE):
@@ -408,101 +399,6 @@ def load_demo_data() -> None:
             )
     db.commit()
     db.close()
-
-def _sanitize_db_name(name: str) -> Optional[str]:
-    """Return a sanitized ``name.db`` or ``None`` if empty after cleaning."""
-
-    base, _ = os.path.splitext(name)
-    safe = re.sub(r"[^A-Za-z0-9_-]", "", base)[:64]
-    if not safe:
-        return None
-    return safe + ".db"
-
-
-def _sanitize_export_name(name: str) -> str:
-    """Return a filename safe for download containing only allowed chars."""
-
-    safe = re.sub(r"[^A-Za-z0-9_-]", "_", name.strip())
-    safe = safe.strip("_") or "download"
-    if not safe.lower().endswith(".db"):
-        safe += ".db"
-    return safe
-
-
-def create_new_db(name: Optional[str] = None) -> str:
-    """Reset the database and return the newly created filename."""
-    nm = _sanitize_db_name(name) if name else 'waybax.db'
-    if nm is None:
-        raise ValueError('Invalid database name.')
-    db_path = os.path.join(app.root_path, nm)
-    if os.path.exists(db_path):
-        os.remove(db_path)
-    app.config['DATABASE'] = db_path
-    init_db()
-    # Databases used to include demo entries from ``data/demo_data.json``.
-    # For production use we initialize an empty database.
-    return nm
-
-if app.config.get('DATABASE') and os.path.exists(app.config['DATABASE']):
-    ensure_schema()
-
-
-@app.before_request
-def _update_display_name() -> None:
-    """Ensure ``session['db_display_name']`` matches the current DB file."""
-    if app.config.get('DATABASE'):
-        actual_name = os.path.basename(app.config['DATABASE'])
-    else:
-        actual_name = '(none)'
-    if session.get('db_display_name') != actual_name:
-        session['db_display_name'] = actual_name
-
-def get_db() -> sqlite3.Connection:
-    """Return a SQLite connection stored on the Flask ``g`` object."""
-
-    if not app.config.get('DATABASE'):
-        raise RuntimeError('No database loaded.')
-    db = getattr(g, '_database', None)
-    if db is None:
-        db = g._database = sqlite3.connect(app.config['DATABASE'])
-        db.row_factory = sqlite3.Row
-        db.create_function('has_tag', 2, _has_tag)
-    return db
-
-@app.teardown_appcontext
-def close_connection(exception: Optional[BaseException]) -> None:
-    """Close the SQLite connection at app teardown and remove it from ``g``."""
-
-    db = g.pop('_database', None)
-    if db is not None:
-        db.close()
-
-def query_db(query: str, args: Union[Tuple, List] = (), one: bool = False) -> Any:
-    """Execute ``query`` and return rows or a single row."""
-
-    cur = get_db().execute(query, args)
-    rv = cur.fetchall()
-    cur.close()
-    return (rv[0] if rv else None) if one else rv
-
-def execute_db(query: str, args: Union[Tuple, List] = ()) -> int:
-    """Execute ``query`` that modifies the DB and return ``lastrowid``."""
-
-    db = get_db()
-    cur = db.execute(query, args)
-    db.commit()
-    return cur.lastrowid
-
-
-def _has_tag(tags: str, tag: str) -> int:
-    """SQLite helper to check if ``tag`` exists in comma-separated ``tags``."""
-
-    tag = tag.strip().lower()
-    for t in tags.split(','):
-        if t.strip().lower() == tag:
-            return 1
-    return 0
-
 
 def _quote_hashtags(expr: str) -> str:
     """Surround bare hashtag terms with quotes for proper tokenization."""
